+++ src/authentication.js
@@ -9,7 +9,6 @@ module.exports.meerkatAuthenticate = async function () {
     try {
         // Request authentication for GitHub. You could also use 'microsoft'
         const session = await vscode.authentication.getSession('github', ['user:email'], { createIfNone: true });
-        console.log(session);
 
         if (session) {
             // Extract the email from the session's account
@@ -43,7 +42,7 @@ module.exports.meerkatAuthenticate = async function () {
                 } else if (token && !token_valid) {
                     vscode.window.showErrorMessage(`Your MeerkatIO free trial has expired. To continue using MeerkatIO Pro features, add a Pro Subscription today!`, "Get MeerkatIO Pro").then((selection) => {
                         if (selection === 'Get MeerkatIO Pro') {
-                            vscode.env.openExternal("https://meerkatio.com/account#subscription");
+                            vscode.env.openExternal("https://buy.stripe.com/5kAcPD5627OO3OoeUV?prefilled_email=" + email);
                         }
                     });
                 }
@@ -81,10 +80,18 @@ module.exports.checkMeerkatAccount = async function () {
         let account_type = data.account_type;
         let token_valid = data.token_valid;
         let email = data.email;
+        let microsoft_teams_webhook = data.microsoft_teams_webhook;
+        let google_chat_webhook = data.google_chat_webhook;
+        let slack_token = data.slack_token;
+        let phone_num = data.phone_num;
 
         context.globalState.update(Constants.ACCOUNT_TYPE, account_type);
         context.globalState.update(Constants.ACCOUNT_VALID, token_valid);
         context.globalState.update(Constants.ACCOUNT_EMAIL, email);
+        context.globalState.update(Constants.ACCOUNT_MICROSOFT_TEAMS_WEBHOOK, microsoft_teams_webhook);
+        context.globalState.update(Constants.ACCOUNT_GOOGLE_CHAT_WEBHOOK, google_chat_webhook);
+        context.globalState.update(Constants.ACCOUNT_SLACK_TOKEN, slack_token);
+        context.globalState.update(Constants.ACCOUNT_PHONE_NUMBER, phone_num);
 
         setExtensionContext(context);
     }