+++ src/constants.js
@@ -6,7 +6,13 @@ module.exports = {
     ACCOUNT_TYPE: "meerkatio_account_type",
     ACCOUNT_VALID: "meerkatio_account_valid",
     ACCOUNT_EMAIL: "meerkatio_email",
+    ACCOUNT_MICROSOFT_TEAMS_WEBHOOK: "meerkatio_teams_webhook",
+    ACCOUNT_GOOGLE_CHAT_WEBHOOK: "meerkatio_google_chat_webhook",
+    ACCOUNT_SLACK_TOKEN: "meerkatio_slack_token",
+    ACCOUNT_PHONE_NUMBER: "meerkatio_phone_num",
 
     // Notification Data
-    NOTIF_HISTORY: "meerkatio_notification_history"
+    NOTIF_METHOD: "meerkatio_notification_method",
+    NOTIF_HISTORY: "meerkatio_notification_history",
+    NOTIF_TRIGGER_DURATION_SECONDS: "meerkatio_notification_trigger_duration_seconds"
 }